precision_farming/requirements.txt
asgiref==3.2.10
cycler==0.10.0
Django==3.0.8
django-restframework==0.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
pandas==1.0.5
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
whitenoise==5.2.0
